coverpage-images: https://github.com/metanorma/metanorma-itu/issues/506
