+++ lib/metanorma/itu/front.rb
@@ -36,9 +36,8 @@ module Metanorma
 
       def title_otherlangs(node, xml)
         node.attributes.each do |k, v|
-          next unless /^(?:annex)?title-(?<lang>.+)$/ =~ k
-          next if lang == @lang
-
+          /^(?:annex)?title-(?<lang>.+)$/ =~ k or next
+          lang == @lang and next
           type = /^annex/.match?(k) ? "annex" : "main"
           xml.title **attr_code(title_attr(type, lang)) do |t|
             t << Metanorma::Utils::asciidoc_sub(v)
@@ -65,9 +64,8 @@ module Metanorma
 
       def other_title_otherlangs(node, xml, type)
         node.attributes.each do |k, v|
-          next unless m = /^#{type}-(?<lang>.+)$/.match(k)
-          next if m[:lang] == @lang
-
+          m = /^#{type}-(?<lang>.+)$/.match(k) or next
+          m[:lang] == @lang and next
           xml.title **attr_code(title_attr(type.sub(/-title/, ""),
                                            m[:lang])) do |t|
             t << Metanorma::Utils::asciidoc_sub(v)
@@ -132,23 +130,17 @@ module Metanorma
       end
 
       def metadata_series(node, xml)
-        node.attr("series") and
-          xml.series **{ type: "main" } do |s|
-            s.title node.attr("series")
-          end
-        node.attr("series1") and
-          xml.series **{ type: "secondary" } do |s|
-            s.title node.attr("series1")
-          end
-        node.attr("series2") and
-          xml.series **{ type: "tertiary" } do |s|
-            s.title node.attr("series2")
-          end
+        { series: "main", series1: "secondary", series2: "tertiary" }
+          .each do |k, v|
+          node.attr(k.to_s) and
+            xml.series **{ type: v } do |s|
+              s.title node.attr(k.to_s)
+            end
+        end
       end
 
       def metadata_recommendationstatus(node, xml)
-        return unless node.attr("recommendation-from")
-
+        node.attr("recommendation-from") or return
         xml.recommendationstatus do |s|
           s.from node.attr("recommendation-from")
           s.to node.attr("recommendation-to") if node.attr("recommendation-to")
@@ -199,6 +191,18 @@ module Metanorma
         end
       end
 
+      def metadata_coverpage_images(node, xml)
+        %w(coverpage-image).each do |n|
+          if a = node.attr(n)
+            xml.send n do |c|
+              a.split(",").each do |x|
+                c.image src: x
+              end
+            end
+          end
+        end
+      end
+
       def metadata_ext(node, xml)
         metadata_doctype(node, xml)
         metadata_subdoctype(node, xml)
@@ -208,6 +212,7 @@ module Metanorma
         metadata_ip_notice(node, xml)
         metadata_techreport(node, xml)
         structured_id(node, xml)
+        metadata_coverpage_images(node, xml)
       end
     end
   end
